调整TASK3旋转角度参数类型为整型，优化相关函数实现

+++ Core/Inc/setParma.h
@@ -30,9 +30,9 @@ float Get_Forward_Kd(void);
 float Get_Forward_Error_Threshold(void);
 
 // Task3相关参数设置函数声明
-void Set_Task3_Rotation_Angle_1(float angle);
-void Set_Task3_Rotation_Angle_2(float angle);
-void Set_Task3_Rotation_Angle_3(float angle);
+void Set_Task3_Rotation_Angle_1(int angle);
+void Set_Task3_Rotation_Angle_2(int angle);
+void Set_Task3_Rotation_Angle_3(int angle);
 void Set_Task3_Steer_Time_C(int time);
 void Set_Task3_Steer_PWM_C(int pwm);
 void Set_Task3_Steer_Time_D(int time);
